new method to make repeals consistent
Because data is cached, there are cases where a resolution valid when the data was gathered is repealed by a later resolution. This method checks that data for later resolutions. If a later repeal is present, it matches to the earlier resolution by name and then marks it as repealed.

Also some minor documentation changes and code simplifications.